requirements.txt
apipkg==1.5
atomicwrites==1.3.0
attrs==19.1.0
certifi==2019.9.11
chardet==3.0.4
colorama==0.4.1
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==2.2.7
django-rest-swagger==2.2.0
djangorestframework==3.10.3
djangorestframework-simplejwt==4.3.0
execnet==1.7.1
idna==2.8
importlib-metadata==0.20
itypes==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
more-itertools==7.2.0
openapi-codec==1.3.2
packaging==19.1
pluggy==0.12.0
psycopg2==2.8.4
py==1.8.0
PyJWT==1.7.1
pyparsing==2.4.2
pytest==5.1.2
pytest-cache==1.0
python-decouple==3.3
pytz==2019.3
requests==2.22.0
simplejson==3.17.0
six==1.12.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.7
virtualenv==16.7.7
virtualenvwrapper-win==1.2.5
wcwidth==0.1.7
zipp==0.6.0
